feat(frontend): finished linking voting to backend

+++ apps/backend/src/shared/entities/cup.entity.ts
@@ -117,7 +117,7 @@ export class Cup {
   @Column({ type: 'character varying', nullable: true })
   state: CupStates;
 
-  @Column({ type: 'integer', nullable: true })
+  @Column({ type: 'float', nullable: true })
   map_rating: number | null;
 
   @Column({ type: 'integer', nullable: true })
